feature(RESULTS.md) - добавил описание запуска
