+++ analysis-of-analytical-storages/RESULTS.md
@@ -2,7 +2,19 @@
 В качестве кандидатов для выбора хранилища для обработки аналитических данных были выбраны ClickHouse и Vertica. Исследоование показало, что ClickHouse быстрее справляется с аналитическими запросами, чем Vertica.
 Для примера сравним скорость выполнения одинаковых запросов в обеих БД.
 
-Описания тестов:
+
+### 🚀 Запуск тестов
+Чтобы поднять тесты надо выполнить следующие команды 
+   (у вас должен быть установлен docker-compose и python3):
+- `pip3 install -r requirements.txt`
+- `docker-compose build`
+- `docker-compose up -d`
+- `python3 test_data/fake_data.py`
+- `python3 test_insert.py`
+- `python3 test_read.py`
+
+
+### 📙 Описания тестов:
 - `select_count` - подсчитать кол-во записей в таблице
   #### запрос: 
       SELECT COUNT(*) FROM view
@@ -50,7 +62,7 @@
       )
       VALUES (%s,%s,%s,%s,%s)'
 
-## Результаты тестов
+### 📘 Результаты тестов
 
 | Тест | ClickHouse, sec | Vertica, sec |
 |---|:---------------:|:------------:|
@@ -65,5 +77,6 @@
 
 
 Как видно по данным, при схожих параметрах, в запросах на получение и вставку данных Clickhouse обгоняет Vertica в разы.
-## Вывод
+
+### Вывод
 Было решено использовать ClickHouse, так как он быстрее справляется с задачами вставки и агрегации.
