+++ analysis-of-analytical-storages/class_speed_test.py
@@ -3,27 +3,29 @@ from abc import ABC, abstractmethod
 
 
 class SpeedTest(ABC):
-    @abstractmethod
+
     def test_insert_data(self, query, data):
-        ...
+        pass
 
     @abstractmethod
     def test_get_data(self, query):
-        ...
+        pass
 
 
-class DBSpeedTest(SpeedTest):
+class ClickHouseSpeedTest(SpeedTest):
     def __init__(self, db_connection):
         self.db = db_connection
+
 
     def test_insert_data(self, query, rows_gen):
         start_time = time.time()
 
         for rows in rows_gen:
             self.db.execute(query, rows)
-            
+
         end_time = time.time()
         return end_time - start_time
+
 
     def test_get_data(self, query):
         start_time = time.time()
@@ -34,7 +36,8 @@ class DBSpeedTest(SpeedTest):
 
 class VerticaSpeedTest(SpeedTest):
     def __init__(self, cursor):
-        self.cursor = cursor
+        self.cursor = cursor.cursor()
+
 
     def test_insert_data(self, query, rows_gen):
         start_time = time.time()
@@ -44,6 +47,7 @@ class VerticaSpeedTest(SpeedTest):
 
         end_time = time.time()
         return end_time - start_time
+
 
     def test_get_data(self, query):
         start_time = time.time()
