--- analysis-of-analytical-storages/clickhouse/README.md
@@ -1,1 +1,0 @@
-Здесь будет информация об исследовании производительности ClickHouse.
