--- analysis-of-analytical-storages/clickhouse/data/node2/users.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0"?>
-<company>
-    <profiles>
-        <default>
-            <max_memory_usage>10000000000</max_memory_usage>
-            <use_uncompressed_cache>0</use_uncompressed_cache>
-            <load_balancing>in_order</load_balancing>
-            <log_queries>1</log_queries>
-        </default>
-    </profiles>
-
-    <users>
-        <default>
-            <password></password>
-            <profile>default</profile>
-            <networks>
-                <ip>::/0</ip>
-            </networks>
-            <quota>default</quota>
-        </default>
-        <admin>
-            <password>123</password>
-            <profile>default</profile>
-            <networks>
-                <ip>::/0</ip>
-            </networks>
-            <quota>default</quota>
-        </admin>
-    </users>
-
-    <quotas>
-        <default>
-            <interval>
-                <duration>3600</duration>
-                <queries>0</queries>
-                <errors>0</errors>
-                <result_rows>0</result_rows>
-                <read_rows>0</read_rows>
-                <execution_time>0</execution_time>
-            </interval>
-        </default>
-    </quotas>
-</company>
