--- analysis-of-analytical-storages/clickhouse/docker-compose.yml
@@ -1,48 +1,0 @@
-version: '3.5'
-
-services:
-  zookeeper:
-    image: zookeeper:3.5
-    container_name: zookeeper
-    hostname: zookeeper
-
-  clickhouse-node1:
-    image: yandex/clickhouse-server:20.4
-    container_name: clickhouse-node1
-    hostname: clickhouse-node1
-    ports:
-      - "8123:8123"
-      - "9000:9000"
-    volumes:
-      - ./data/node1:/etc/clickhouse-server
-    depends_on:
-      - zookeeper
-
-  clickhouse-node2:
-    image: yandex/clickhouse-server:20.4
-    container_name: clickhouse-node2
-    hostname: clickhouse-node2
-    volumes:
-      - ./data/node2:/etc/clickhouse-server
-    depends_on:
-      - zookeeper
-
-  clickhouse-node3:
-    image: yandex/clickhouse-server:20.4
-    container_name: clickhouse-node3
-    hostname: clickhouse-node3
-    ports:
-      - "8124:8123"
-    volumes:
-      - ./data/node3:/etc/clickhouse-server
-    depends_on:
-      - zookeeper
-
-  clickhouse-node4:
-    image: yandex/clickhouse-server:20.4
-    container_name: clickhouse-node4
-    hostname: clickhouse-node4
-    volumes:
-      - ./data/node4:/etc/clickhouse-server
-    depends_on:
-      - zookeeper
