+++ analysis-of-analytical-storages/db/clickhouse/data_scheme.py
