+++ analysis-of-analytical-storages/db/vertica/data_scheme.py
