+++ analysis-of-analytical-storages/db/vertica/vertica_client.py
