+++ analysis-of-analytical-storages/docker-compose.yml
