+++ analysis-of-analytical-storages/requirements.txt
@@ -1,0 +1,2 @@
+clickhouse-driver==0.2.5
+vertica-python==1.3.0
