+++ analysis-of-analytical-storages/test_insert.py
@@ -2,14 +2,14 @@ from db.clickhouse.clickhouse_client import ch_client
 from db.clickhouse.data_scheme import flush_ch, init_ch
 from db.vertica.data_scheme import flush_vertica, init_vertica
 from db.vertica.vertica_client import vt_client
-from speed_test import DBSpeedTest, VerticaSpeedTest
+from class_speed_test import ClickHouseSpeedTest, VerticaSpeedTest
 from test_data.fake_data import data_generator
 
-ch_speed_test = DBSpeedTest(ch_client)
+ch_speed_test = ClickHouseSpeedTest(ch_client)
 flush_ch(ch_client)
 init_ch(ch_client)
 
-vertica_speed_test = VerticaSpeedTest(vt_client.cursor())
+vertica_speed_test = VerticaSpeedTest(vt_client)
 flush_vertica(vt_client.cursor())
 init_vertica(vt_client.cursor())
 
@@ -27,4 +27,6 @@ def test_insert():
     )
     print(f"Write Vertica - {b}sec")
 
-test_insert()
+
+if __name__ == "__main__":
+    test_insert()
