+++ analysis-of-analytical-storages/test_read.py
@@ -1,22 +1,23 @@
 from db.clickhouse.clickhouse_client import ch_client
 from db.vertica.vertica_client import vt_client
-from speed_test import DBSpeedTest, VerticaSpeedTest
+from class_speed_test import ClickHouseSpeedTest, VerticaSpeedTest
 from test_data.fake_data import data_generator
 
-ch_speed_test = DBSpeedTest(ch_client)
-vertica_speed_test = VerticaSpeedTest(vt_client.cursor())
+ch_speed_test = ClickHouseSpeedTest(ch_client)
+vertica_speed_test = VerticaSpeedTest(vt_client)
 
 
+def execute_query(times, query, speed_test):
+    t = 0
+    for _ in range(times):
+        t += speed_test.test_get_data(query)
+    return t
 def test_read_query(query, times=3):
     print(f'*** {query} ***')
-    t = 0
-    for _ in range(times):
-        t += ch_speed_test.test_get_data(query)
+    t = execute_query(times, query, ch_speed_test)
     print(f"CLICK HOUSE - {t/times:.4f} sec")
 
-    t = 0
-    for _ in range(times):
-        t += vertica_speed_test.test_get_data(query)
+    t = execute_query(times, query, vertica_speed_test)
     print(f"VERTICA - {t/times:.4f} sec")
 
     print("")
@@ -48,4 +49,5 @@ def test_read():
         test_read_query(query)
 
 
-test_read()
+if __name__ == "__main__":
+    test_read()
