--- analysis-of-analytical-storages/vokson/read.log
@@ -1,39 +1,0 @@
-*** SELECT COUNT(*) FROM views ***
-CLICK HOUSE - 0.0031 sec
-VERTICA - 0.0117 sec
-
-*** SELECT count(DISTINCT movie_id) FROM views ***
-CLICK HOUSE - 0.1761 sec
-VERTICA - 0.4499 sec
-
-*** SELECT count(DISTINCT user_id) FROM views ***
-CLICK HOUSE - 0.3795 sec
-VERTICA - 1.0505 sec
-
-*** SELECT user_id, count(distinct movie_id) FROM views GROUP by user_id ***
-CLICK HOUSE - 2.8090 sec
-VERTICA - 11.3791 sec
-
-*** 
-        SELECT 
-            user_id, 
-            sum(viewed_frame),
-            max(viewed_frame) 
-        FROM views
-        GROUP by user_id
-         ***
-CLICK HOUSE - 1.1621 sec
-VERTICA - 4.1204 sec
-
-*** 
-        SELECT 
-            user_id, 
-            sum(viewed_frame),
-            max(viewed_frame) 
-        FROM views
-        WHERE event_time > '2021-04-13 23:09:02'
-        GROUP by user_id
-         ***
-CLICK HOUSE - 0.8180 sec
-VERTICA - 3.5423 sec
-
